+++ packages/svelte-query-persist-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tanstack/svelte-query-persist-client
 
+## 6.0.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.13
+  - @tanstack/svelte-query@6.0.12
+
 ## 6.0.13
 
 ### Patch Changes